+++ Chimp/PebcakException.cs
@@ -1,3 +1,3 @@
 namespace Chimp;
 
-public class PebcakException(string message) : Exception(message);
+public class PebcakException(string message, Dictionary<string, object>? args = null) : LocalizedException(message, args);