+++ Chimp/TimeInterval.cs
@@ -1,48 +1,53 @@
+using System.Globalization;
 using System.Text.RegularExpressions;
 
 namespace Chimp;
 
 public class TimeInterval
 {
-    private readonly string _rawInterval;
     private DateTime _baseDate;
-    private readonly Localizer _localizer;
 
     public bool InputContainsWeekDay { get; private set; }
     public DateTime Start { get; private set; }
     public DateTime End { get; private set; }
 
 
-    public TimeInterval(string intervalString, DateTime baseDate, Localizer localizer)
+    public TimeInterval(string intervalString, DateTime baseDate, CultureInfo culture)
     {
         if (baseDate.Kind == DateTimeKind.Unspecified) throw new ApplicationException($"{nameof(baseDate)} cannot have DateTimeKind.Unspecified");
 
-        _rawInterval = intervalString;
         _baseDate = baseDate;
-        _localizer = localizer;
-
-        ProcessLocalizedDayPrefix(intervalString);
+        if (!TryProcessLocalizedDayPrefix(intervalString, culture)
+            // English dayprefix fallback
+            && !TryProcessLocalizedDayPrefix(intervalString, CultureInfo.GetCultureInfo("en")))
+            throw new PebcakException("could not parse the weekdayprefix: use 'mo:', 'tu:', 'we:', 'th:' or 'fr:'");
+        
         if (InputContainsWeekDay) intervalString = intervalString[3..];
 
         if (!TryParseStartTimePlusMinutes(intervalString) && !ParseStartTimePlusEndTime(intervalString))
-            throw new PebcakException($"could not parse {nameof(intervalString)}");
+            throw new PebcakException("cannot parse timeSpec '{TimeSpec}'", new() {{"TimeSpec", intervalString}});
     }
 
-    private void ProcessLocalizedDayPrefix(string input)
+    /// <returns>false if dayprefix was invalid</returns>
+    private bool TryProcessLocalizedDayPrefix(string input, CultureInfo culture)
     {
         var dayPrefixMatch = Regex.Match(input, @"^(?<Day>[a-z]{2}):");
         InputContainsWeekDay = dayPrefixMatch.Success;
-        if (!InputContainsWeekDay) return;
+        if (!InputContainsWeekDay) return true;
 
-        _baseDate = Util.GetFirstDayOfWeek(_baseDate);
+        var newBaseDate = Util.GetFirstDayOfWeek(_baseDate);
         for (var i = 0; i < 7; i++)
         {
             // FIXME we assume here that the first two letters of the weekday are unique for the users' language
-            if (string.Equals(dayPrefixMatch.Groups["Day"].Value, _localizer.GetWeekDay(_baseDate)[..2], StringComparison.OrdinalIgnoreCase)) return;
-            _baseDate = _baseDate.AddDays(1);
+            if (string.Equals(dayPrefixMatch.Groups["Day"].Value, string.Create(culture, $"{newBaseDate:dddd}")[..2], StringComparison.OrdinalIgnoreCase))
+            {
+                _baseDate = newBaseDate;
+                return true;
+            }
+            newBaseDate = newBaseDate.AddDays(1);
         }
-
-        throw new PebcakException("could not parse weekdayprefix: make sure to use the language known in timechimp, e.g. use 'ma:' if you are Dutch or 'mo:' if you are English");
+        
+        return false;
     }
 
     private bool TryParseStartTimePlusMinutes(string interval)
@@ -56,9 +61,9 @@ public class TimeInterval
             .AddMinutes(ParseMinute(matchRelative.Groups["StartMinute"].Value));
 
         var totalMinutes = int.Parse(matchRelative.Groups["TotalMinutes"].Value);
-        if (totalMinutes > 10 * 60) throw new PebcakException("to help preventing input mistakes, a time interval is not allowed not exceed 10 hours");
+        if (totalMinutes > 10 * 60) throw new PebcakException("timeSpec '{TimeSpec}' issue: to help preventing input mistakes, a time interval is not allowed not exceed 10 hours", new(){{"TimeSpec", interval}});
         End = Start.AddMinutes(totalMinutes);
-        if (End.Date != Start.Date) throw new PebcakException($"{nameof(interval)} start and end must be on the same day");
+        if (End.Date != Start.Date) throw new PebcakException("timeSpec '{TimeSpec}' issue: start and end must be on the same day", new(){{"TimeSpec", interval}});
 
         return true;
     }
@@ -77,8 +82,8 @@ public class TimeInterval
             .AddHours(ParseHour(matchEndtime.Groups["EndHour"].Value))
             .AddMinutes(ParseMinute(matchEndtime.Groups["EndMinute"].Value));
 
-        if (End <= Start) throw new PebcakException($"{nameof(interval)} end time should be after start time");
-        if (End.Date != Start.Date) throw new PebcakException($"{nameof(interval)} start and end must be on the same day");
+        if (End <= Start) throw new PebcakException("timeSpec '{TimeSpec}' issue: end time should be after start time");
+        if (End.Date != Start.Date) throw new PebcakException("timeSpec '{TimeSpec}' issue: start and end must be on the same day", new(){{"TimeSpec", interval}});
 
         return true;
     }
@@ -86,13 +91,13 @@ public class TimeInterval
     private static int ParseHour(string strHour)
     {
         var hour = int.Parse(strHour);
-        return hour is < 0 or > 23 ? throw new PebcakException("hour is out of range") : hour;
+        return hour is < 0 or > 23 ? throw new PebcakException("hour '{Hour}' is invalid", new(){{"Hour", strHour}}) : hour;
     }
 
     private static int ParseMinute(string strMinute)
     {
         if (string.IsNullOrWhiteSpace(strMinute)) return 0;
         var minute = int.Parse(strMinute);
-        return minute is < 0 or > 59 ? throw new PebcakException("minute is out of range") : minute;
+        return minute is < 0 or > 59 ? throw new PebcakException("minute '{Minute}' is invalid", new(){{"Minute", strMinute}}) : minute;
     }
 }