+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Release v1.2 (30-dec-2023)
+
+- localized and improved most messages
+- English fallback for weekday prefix
+- no known issues
+
 ## Release v1.1 (29-dec-2023)
 
 - added support for adding tags